% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow.R
\name{follow_.default}
\alias{follow_.default}
\title{follow_.default}
\usage{
\method{follow_}{default}(users)
}
\arguments{
\item{users}{Screen names or user IDs of users to follow.}
}
\value{
API response objects (invisibly)
}
\description{
Follows users with random sleep timer between calls.
}
