% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow.R
\name{follow_}
\alias{follow_}
\title{follow_}
\usage{
follow_(users)
}
\arguments{
\item{users}{Screen names or user IDs of users to follow.}
}
\value{
API response objects (invisibly)
}
\description{
Follows users with random sleep timer between calls.
}
