% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_retweets.R
\name{no_retweets}
\alias{no_retweets}
\title{no_retweets}
\usage{
no_retweets(n = 1000L)
}
\arguments{
\item{n}{Number of most recent users to mute retweets.}
}
\value{
API response object.
}
\description{
Mutes retweets from most recent n users
}
