% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/follow.R
\name{regen_token}
\alias{regen_token}
\title{regen_token}
\usage{
regen_token(browse = TRUE)
}
\arguments{
\item{browse}{Logical indicating whether or not to open the web
browser to apps.twitter.com/app/}
}
\value{
Token (invisible).
}
\description{
Regenerates oauth token.
}
