% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickrtweet.R
\name{twitter_stats}
\alias{twitter_stats}
\title{twitter statistics}
\usage{
twitter_stats()
}
\value{
Invisibly returns a list of friends, friends data, followers, 
  and followers data
}
\description{
Displays Twitter statistics of home user and user network and returns 
  friends and followers data.
}
