% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfollowback.R
\name{unfollowback}
\alias{unfollowback}
\title{unfollowback}
\usage{
unfollowback(keep = NULL, ...)
}
\arguments{
\item{keep}{Character vector of screen names or IDs to not unfollow.
These are probably popular accounts that you don't expect reciprocity
from. Default is NULL. If "../data/keepers.rds" exists, it will be
read in and used regardless of whether this value is NULL.}

\item{...}{Passed to \code{\link{unfollowback.default}}}
}
\value{
List where each element corresponds to an API response object.
  Elements should be of status 200 if unfollow request was successful.
}
\description{
Removes friends who don't follow back (with exceptions)
}
