% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_retweets.R
\name{no_retweets_}
\alias{no_retweets_}
\title{no_retweets_}
\usage{
no_retweets_(n = 1000L, refresh = TRUE)
}
\arguments{
\item{n}{Number of most recent users to mute retweets.}

\item{refresh}{Logical indicating whether to refresh friends list.}
}
\value{
API response object.
}
\description{
Mutes retweets from most recent n users
}
